% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecofun-package.R
\docType{package}
\name{ecofun-package}
\alias{ecofun}
\alias{ecofun-package}
\title{ecofun: A Collection Of Useful Functions For Ecological Analysis}
\description{
A collection of useful functions for ecological analysis
}
\author{
\strong{Maintainer}: George Roff \email{george.roff@csiro.au} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
