% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla2.R
\name{inla2}
\alias{inla2}
\title{Wrapper to simplify INLA models}
\usage{
inla2(
  data,
  formula,
  newdat,
  family = "gaussian",
  compute_dic = TRUE,
  compute_waic = TRUE,
  compute_marginals = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{data}{df for inla model}

\item{formula}{specified formula (no ranef yet)}

\item{newdat}{newdata for predictors as list}

\item{family}{default gaussian}

\item{compute_dic}{compute dic, logical}

\item{compute_waic}{compute waic, logical}

\item{compute_marginals}{compute waic, logical}

\item{verbose}{silence inla}

\item{...}{passes functions}
}
\description{
Function to wrap inla models iwth predictors
}
