% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\title{Wrapper to predict from INLA models}
\usage{
predict(input, ...)
}
\arguments{
\item{input}{input from ecofun::inla}

\item{...}{passes functions}
}
\description{
Function to wrap inla models with predictors
}
