% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla2.R
\name{inla2}
\alias{inla2}
\title{Wrapper to simplify INLA models}
\usage{
inla2(
  formula,
  data,
  newdat,
  response_log = FALSE,
  seed = NULL,
  family = "gaussian",
  compute_dic = FALSE,
  compute_waic = FALSE,
  compute_cpo = FALSE,
  compute_marginals = FALSE
)
}
\arguments{
\item{formula}{Specified formula (no random effects yet)}

\item{data}{Data frame for INLA model}

\item{newdat}{New data for predictors as a list}

\item{family}{Family distribution for INLA, default is "gaussian"}

\item{compute_dic}{Compute DIC, logical}

\item{compute_waic}{Compute WAIC, logical}

\item{compute_cpo}{Compute CPO, logical}

\item{compute_marginals}{Compute marginals, logical}

\item{...}{Additional parameters passed to INLA}
}
\description{
Function to wrap INLA models with predictors
}
\examples{
# Set seed for reproducibility
set.seed(123)
}
